+++ src/extension.ts
@@ -50,49 +50,12 @@ export function activate(context: vscode.ExtensionContext) {
 		createdClassProvider.loadDocument(editor.document, doOverWrite);
 		registeredAliasProvider.loadDocument(editor.document, doOverWrite);
 		if(!config.get('neverShowSuccessfulInformation')){
-			vscode.window.showInformationMessage('Successfully load document!');
+			vscode.window.showInformationMessage('Successfully activate!');
 		}
 	};
-
-	const loadSelection = () => {
-		if (editor === undefined) {
-			vscode.window.showInformationMessage('Failed to deap-supporter.loadSelection because activeTextEditor is undefined.');
-			return;
-		}
-		const doOverWrite: boolean|undefined = config.get('alwaysOverwrite');
-		let selection: vscode.Selection = editor.selection;
-		if (selection.isEmpty) {
-			vscode.window.showInformationMessage('failed to deap-supporter.loadSelection because of no selection');
-			return;
-		}
-		let startRow: number = selection.active.line;
-		let lines: string[] = editor.document.getText(selection).split('\n');
-		lines.forEach((line, index) => {
-			createdClassProvider.loadLine(line, startRow + index, doOverWrite);
-			registeredAliasProvider.loadLine(line, startRow + index, doOverWrite);
-		});
-		if(!config.get('neverShowSuccessfulInformation')){
-			vscode.window.showInformationMessage('Successfully load selection!');
-		}
-	};
-
-	const reloadDocument = () => {
-		if (editor === undefined) {
-			vscode.window.showInformationMessage('Failed to deap-supporter. loadDocument because activeTextEditor is undefined.');
-			return;
-		}
-		createdClassProvider.reloadDocument(editor.document);
-		registeredAliasProvider.reloadDocument(editor.document);
-		if(!config.get('neverShowSuccessfulInformation')){
-			vscode.window.showInformationMessage('Successfully reload document!');
-		}
-	};
-
 
 	context.subscriptions.push(
-		vscode.commands.registerCommand('deap-supporter.loadDocument', loadDocument),
-		vscode.commands.registerCommand('deap-supporter.loadSelection', loadSelection),
-		vscode.commands.registerCommand('deap-supporter.reloadDocument', reloadDocument)
+		vscode.commands.registerCommand('deap-supporter.activate', loadDocument)
 	);
 }