+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to the "deap-supporter-ts" extension will be documented in t
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [2.0.0] - 2021-06-21
+
+### changed
+
+- All of the commands exiting before have been abolished.
+  - Because there may be no need to reload current document.
+    - Moreover, it's a pain in the ass to load or reload.
+  - Instead, add the command to activate this extension.
+
+### added
+
+- added the command to activate this extension
+  - once you activate, there is no need to load current document.
+  - automatically load and add your custom classes or aliases!
+
+
 ## [1.0.1] - 2021-03-19
 
 ### Fixed